move unit test function to unit tests
